+++ 1.18/Common/src/main/java/fuzs/blockrunner/client/BlockRunnerClient.java
@@ -1,0 +1,17 @@
+package fuzs.blockrunner.client;
+
+import fuzs.blockrunner.client.handler.BlockSpeedTooltipHandler;
+import fuzs.puzzleslib.api.client.core.v1.ClientModConstructor;
+import fuzs.puzzleslib.api.client.event.v1.ItemTooltipCallback;
+
+public class BlockRunnerClient implements ClientModConstructor {
+
+    @Override
+    public void onConstructMod() {
+        registerHandlers();
+    }
+
+    private static void registerHandlers() {
+        ItemTooltipCallback.EVENT.register(BlockSpeedTooltipHandler::onItemTooltip);
+    }
+}