+++ 1.18/CHANGELOG.md
@@ -3,7 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v3.0.0-1.18.2] - 2023-06-25
+## [v3.1.0-1.18.2] - 2023-08-12
 - Ported to Minecraft 1.18.2
+
+## [v3.0.2-1.18.2] - 2022-07-30
+### Fixed
+- Ported fix for slabs not working correctly from 1.19 version
+
+## [v3.0.1-1.18.2] - 2022-03-15
+### Changed
+- Slightly lowered default speed values
+
+## [v3.0.0-1.18.2] - 2022-03-14
+- Reworked for Minecraft 1.18
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/